+++ src/Sylius/Bundle/ShippingBundle/DependencyInjection/Compiler/RegisterRuleCheckersPass.php
@@ -24,11 +24,11 @@ class RegisterRuleCheckersPass implements CompilerPassInterface
 {
     public function process(ContainerBuilder $container)
     {
-        if (!$container->hasDefinition('sylius.shipping_rule_checker_registry')) {
+        if (!$container->hasDefinition('sylius.registry.shipping_rule_checker')) {
             return;
         }
 
-        $registry = $container->getDefinition('sylius.shipping_rule_checker_registry');
+        $registry = $container->getDefinition('sylius.registry.shipping_rule_checker');
         $checkers = array();
 
         foreach ($container->findTaggedServiceIds('sylius.shipping_rule_checker') as $id => $attributes) {
@@ -37,7 +37,7 @@ class RegisterRuleCheckersPass implements CompilerPassInterface
             }
             $checkers[$attributes[0]['type']] = $attributes[0]['label'];
 
-            $registry->addMethodCall('registerChecker', array($attributes[0]['type'], new Reference($id)));
+            $registry->addMethodCall('register', array($attributes[0]['type'], new Reference($id)));
         }
 
         $container->setParameter('sylius.shipping_rules', $checkers);