+++ src/Sylius/Bundle/ShippingBundle/Form/EventListener/BuildRuleFormSubscriber.php
@@ -11,7 +11,7 @@
 
 namespace Sylius\Bundle\ShippingBundle\Form\EventListener;
 
-use Sylius\Component\Shipping\Checker\Registry\RuleCheckerRegistryInterface;
+use Sylius\Component\Registry\ServiceRegistryInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
@@ -27,7 +27,7 @@ use Symfony\Component\Form\FormInterface;
 class BuildRuleFormSubscriber implements EventSubscriberInterface
 {
     /**
-     * @var RuleCheckerRegistryInterface
+     * @var ServiceRegistryInterface
      */
     private $checkerRegistry;
 
@@ -36,7 +36,11 @@ class BuildRuleFormSubscriber implements EventSubscriberInterface
      */
     private $factory;
 
-    public function __construct(RuleCheckerRegistryInterface $checkerRegistry, FormFactoryInterface $factory)
+    /**
+     * @param ServiceRegistryInterface $checkerRegistry
+     * @param FormFactoryInterface     $factory
+     */
+    public function __construct(ServiceRegistryInterface $checkerRegistry, FormFactoryInterface $factory)
     {
         $this->checkerRegistry = $checkerRegistry;
         $this->factory = $factory;
@@ -85,12 +89,12 @@ class BuildRuleFormSubscriber implements EventSubscriberInterface
 
     /**
      * @param FormInterface $form
-     * @param $ruleType
+     * @param string        $ruleType
      * @param array         $data
      */
     protected function addConfigurationFields(FormInterface $form, $ruleType, array $data = array())
     {
-        $checker = $this->checkerRegistry->getChecker($ruleType);
+        $checker = $this->checkerRegistry->get($ruleType);
         $configurationField = $this->factory->createNamed('configuration', $checker->getConfigurationFormType(), $data, array('auto_initialize' => false));
 
         $form->add($configurationField);