+++ src/Sylius/Bundle/ShippingBundle/Form/EventListener/BuildShippingMethodFormSubscriber.php
@@ -11,7 +11,7 @@
 
 namespace Sylius\Bundle\ShippingBundle\Form\EventListener;
 
-use Sylius\Component\Shipping\Calculator\Registry\CalculatorRegistryInterface;
+use Sylius\Component\Registry\ServiceRegistryInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Form\FormEvent;
 use Symfony\Component\Form\FormEvents;
@@ -29,7 +29,7 @@ class BuildShippingMethodFormSubscriber implements EventSubscriberInterface
     /**
      * It hold registry of all calculators.
      *
-     * @var CalculatorRegistryInterface
+     * @var ServiceRegistryInterface
      */
     private $calculatorRegistry;
 
@@ -43,10 +43,10 @@ class BuildShippingMethodFormSubscriber implements EventSubscriberInterface
     /**
      * Constructor.
      *
-     * @param CalculatorRegistryInterface $calculatorRegistry
-     * @param FormFactoryInterface        $factory
+     * @param ServiceRegistryInterface $calculatorRegistry
+     * @param FormFactoryInterface     $factory
      */
-    public function __construct(CalculatorRegistryInterface $calculatorRegistry, FormFactoryInterface $factory)
+    public function __construct(ServiceRegistryInterface $calculatorRegistry, FormFactoryInterface $factory)
     {
         $this->calculatorRegistry = $calculatorRegistry;
         $this->factory = $factory;
@@ -104,7 +104,7 @@ class BuildShippingMethodFormSubscriber implements EventSubscriberInterface
      */
     protected function addConfigurationFields(FormInterface $form, $calculatorName, array $data = array())
     {
-        $calculator = $this->calculatorRegistry->getCalculator($calculatorName);
+        $calculator = $this->calculatorRegistry->get($calculatorName);
 
         if (!$calculator->isConfigurable()) {
             return;