+++ src/Sylius/Bundle/ShippingBundle/CHANGELOG.md
@@ -39,3 +39,16 @@ CHANGELOG
 * Remove the ``sylius_shipping`` prefix from services and models.
   Use simple ``sylius`` instead.
 * Tag ``sylius_shipping.calculator`` was renamed to ``sylius.shipping_calculator``.
+
+### v0.16.0
+
+* Change service sylius.shipping_calculator_registry to sylius.registry.shipping_calculator
+  and sylius.shipping_rule_checker_registry to sylius.registry.shipping_rule_checker
+* Change service id for shipping rule checker registry to sylius.registry.shipping_rule_checker in RegisterRuleCheckersPass,
+  RegisterRuleCheckersPassSpec
+* Change id of service to sylius.registry.shipping_calculator in RegisterCalculatorsPassSpec, RegisterCalculatorsPass
+* Replace custom RuleCheckerRegistry with ServiceRegistry in BuildRuleFormSubscriber, RuleType, ShippingMethodType,
+  BuildRuleFormSubscriberSpec, RuleTypeSpec, ShippingMethodTypeSpec
+* Replace custom CalculatorRegistry with ServiceRegistry in BuildShippingMethodFormSubscriber, ShippingMethodChoiceType, ShippingMethodType,
+  BuildShippingMethodFormSubscriberSpec, ShippingMethodChoiceTypeSpec
+